Add auto-sorting toggle and improve undo feature
Users can enable auto-sorting for a folder (confirmation popup with an option to disable), it will watch for new files in selected folder and sort them automatically.

Added undo history, works in a similar way to git commits. And for autosorting, if after adding a file, another is added within 2secs, it is counted as one batch of sort. Each batch can be undone alogn with manual folder sorts.